requirements.txt
langchain==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langgraph>=0.2.20
langgraph-checkpoint
python-dotenv==1.0.1
grandalf==0.8